DDF-3605 Refactored CswActionTransformerProvider into an interface
