+++ catalog/spatial/csw/spatial-csw-endpoint/src/main/java/org/codice/ddf/spatial/ogc/csw/catalog/endpoint/transformer/CswActionTransformerProvider.java
@@ -13,71 +13,17 @@
  */
 package org.codice.ddf.spatial.ogc.csw.catalog.endpoint.transformer;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import javax.annotation.Nullable;
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * <b> This code is experimental. While this interface is functional and tested, it may change or be
  * removed in a future version of the library. </b>
  *
- * <p>Manages a reference list of {@link CswActionTransformer}'s by mapping them to the QName's they
- * apply to.
+ * <p>Manages a reference list of {@link CswActionTransformer}'s by mapping them to the typenames
+ * they apply to.
  */
-public class CswActionTransformerProvider {
+public interface CswActionTransformerProvider {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(CswActionTransformerProvider.class);
-
-  private final Map<String, CswActionTransformer> transformerMap = new HashMap<>();
-
-  public synchronized void bind(CswActionTransformer cswActionTransformer) {
-    if (cswActionTransformer == null) {
-      return;
-    }
-
-    List<String> typeNames = cswActionTransformer.getTypeNames();
-    if (CollectionUtils.isEmpty(typeNames)) {
-      LOGGER.warn("Unable to bind a CswActionTransformer with no typenames");
-      return;
-    }
-
-    for (String typeName : typeNames) {
-      if (transformerMap.containsKey(typeName)) {
-        LOGGER.warn(
-            "Unable to bind a CswActionTransformer: a transformer with typename [{}] has already been added",
-            typeName);
-        continue;
-      }
-      transformerMap.put(typeName, cswActionTransformer);
-    }
-  }
-
-  public synchronized void unbind(CswActionTransformer cswActionTransformer) {
-    if (cswActionTransformer == null) {
-      return;
-    }
-
-    List<String> typeNames = cswActionTransformer.getTypeNames();
-    if (CollectionUtils.isEmpty(typeNames)) {
-      return;
-    }
-
-    for (String typeName : typeNames) {
-      transformerMap.remove(typeName);
-    }
-  }
-
-  public synchronized Optional<CswActionTransformer> getTransformer(@Nullable String typeName) {
-    if (StringUtils.isEmpty(typeName)) {
-      return Optional.empty();
-    }
-
-    return Optional.ofNullable(transformerMap.get(typeName));
-  }
+  Optional<CswActionTransformer> getTransformer(@Nullable String typeName);
 }
